refactor: DTO 위치 및 이름 수정
